--- app-project-old/.gitignore
@@ -1,10 +1,0 @@
-# Dependency directory
-node_modules
-
-# Debug log from npm
-npm-debug.log
-
-# Environment Variables should NEVER be published
-.env
-
-utils
