--- app-project-old/public/stylesheets/style.css
@@ -1,53 +1,0 @@
-body {
-  background-color: #F9F9F9; }
-
-.container {
-  margin-top: 5%; }
-
-.card-header {
-  padding: 5px 15px; }
-
-.profile-img {
-  width: 96px;
-  height: 96px;
-  margin: 0 auto 10px;
-  display: block;
-  -moz-border-radius: 50%;
-  -webkit-border-radius: 50%;
-  border-radius: 50%; }
-
-.forget-pwd > a {
-  color: #dc3545;
-  font-weight: 500; }
-
-.forget-password .panel-default {
-  padding: 31%;
-  margin-top: -27%; }
-
-.forget-password .panel-body {
-  padding: 15%;
-  margin-bottom: -50%;
-  background: #fff;
-  border-radius: 5px;
-  -webkit-box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19); }
-
-img {
-  width: 40%;
-  margin-bottom: 10%; }
-
-.btnForget {
-  background: #c0392b;
-  border: none; }
-
-.forget-password .dropdown {
-  width: 100%;
-  border: 1px solid #ced4da;
-  border-radius: .25rem; }
-
-.forget-password .dropdown button {
-  width: 100%; }
-
-.forget-password .dropdown ul {
-  width: 100%; }
-
-/*# sourceMappingURL=style.css.map */
