--- app-project-old/routes/index.js
@@ -1,10 +1,0 @@
-const express = require('express');
-const router = express.Router();
-
-/* GET home page */
-router.get('/', (req, res, next) => {
-  // res.render('index');
-  res.redirect('/login')
-});
-
-module.exports = router;
