--- app-project-old/views/error.hbs
@@ -1,2 +1,0 @@
-<h1>Error!</h1>
-<h2>There was an error on the server! Check the console!</h2>
