--- app-project-old/views/not-found.hbs
@@ -1,2 +1,0 @@
-<h1>404 Error</h1>
-<h2>This route does not exist on the app! Check if everything is ok!</h2>
