+++ package.json
@@ -1,21 +1,26 @@
 {
-  "name": "app-dopt-me",
+  "name": "passport-example",
   "version": "0.0.0",
   "private": true,
   "scripts": {
     "start": "node ./bin/www",
-    "dev": "DEBUG=app-dopt-me:* nodemon ./bin/www",
+    "dev": "DEBUG=passport-example:* nodemon ./bin/www",
     "dev-windows": "nodemon ./bin/www"
   },
   "dependencies": {
     "body-parser": "^1.18.3",
+    "connect-ensure-login": "^0.1.1",
     "cookie-parser": "^1.4.3",
     "dotenv": "^6.0.0",
     "express": "^4.16.3",
+    "express-session": "^1.16.2",
     "hbs": "^4.0.1",
+    "localtunnel": "^1.9.2",
     "mongoose": "^5.2.10",
     "morgan": "^1.9.0",
     "node-sass-middleware": "^0.11.0",
+    "passport": "^0.4.0",
+    "passport-local-mongoose": "^5.0.1",
     "serve-favicon": "^2.5.0"
   },
   "devDependencies": {
