+++ public/stylesheets/style.css
@@ -1,41 +1,53 @@
 body {
-  padding-top: 60px;
-  padding-left: 0px;
-  padding-right: 0px;
-  font: 14px 'Lucida Grande', Helvetica, Arial, sans-serif;
-  background-color: #4ABDAC;
-  color: white;
-  height: auto; }
+  background-color: #F9F9F9; }
 
-a {
-  color: #4ABDAC; }
+.container {
+  margin-top: 5%; }
 
-.btn-primary {
-  background-color: #fc4a1a;
+.card-header {
+  padding: 5px 15px; }
+
+.profile-img {
+  width: 96px;
+  height: 96px;
+  margin: 0 auto 10px;
+  display: block;
+  -moz-border-radius: 50%;
+  -webkit-border-radius: 50%;
+  border-radius: 50%; }
+
+.forget-pwd > a {
+  color: #dc3545;
+  font-weight: 500; }
+
+.forget-password .panel-default {
+  padding: 31%;
+  margin-top: -27%; }
+
+.forget-password .panel-body {
+  padding: 15%;
+  margin-bottom: -50%;
+  background: #fff;
+  border-radius: 5px;
+  -webkit-box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19); }
+
+img {
+  width: 40%;
+  margin-bottom: 10%; }
+
+.btnForget {
+  background: #c0392b;
   border: none; }
 
-.navbar-custom {
-  background-color: #FC4A1A; }
+.forget-password .dropdown {
+  width: 100%;
+  border: 1px solid #ced4da;
+  border-radius: .25rem; }
 
-.carousel-inner {
-  width: 100vw;
-  height: 50vh; }
+.forget-password .dropdown button {
+  width: 100%; }
 
-.carousel-item {
-  margin: 0 auto; }
-
-.card-body {
-  background-color: #f0f0f0;
-  color: #595959; }
-
-.card-deck {
-  padding: 30px 0px; }
-
-.card-footer {
-  background-color: #f0f0f0;
-  color: #4ABDAC; }
-
-footer {
-  background-color: #c24523; }
+.forget-password .dropdown ul {
+  width: 100%; }
 
 /*# sourceMappingURL=style.css.map */
