+++ public/stylesheets/style.scss
@@ -1,46 +1,56 @@
-body {
-  padding-top: 60px;
-  padding-left: 0px;
-  padding-right: 0px;
-  font: 14px 'Lucida Grande', Helvetica, Arial, sans-serif;
-  background-color: #4ABDAC;
-  color: white;
-  height: auto;
+body { 
+		background-color: #F9F9F9;
 }
 
-a {
-  color: #4ABDAC;
+.container{
+	margin-top: 5%;
 }
 
-.btn-primary {
-  background-color: #fc4a1a;
-  border: none;
+.card-header {
+    padding: 5px 15px;
 }
 
-.navbar-custom {
-  background-color: #FC4A1A;
-}
-.carousel-inner{
- width: 100vw;
-  height: 50vh;
-}
-.carousel-item{
-  margin: 0 auto;
-}
-
-.card-body{
-  background-color: #f0f0f0 ;
-color: #595959;
+.profile-img {
+    width: 96px;
+	height: 96px;
+    margin: 0 auto 10px;
+    display: block;
+    -moz-border-radius: 50%;
+    -webkit-border-radius: 50%;
+    border-radius: 50%;
 }
 
-.card-deck{
-  padding: 30px 0px;
+.forget-pwd > a{
+    color: #dc3545;
+    font-weight: 500;
 }
-.card-footer{
-  background-color: #f0f0f0 ;
-  color:  #4ABDAC;
+.forget-password .panel-default{
+    padding: 31%;
+    margin-top: -27%;
 }
-
-footer{
-  background-color: #c24523;
+.forget-password .panel-body{
+    padding: 15%;
+    margin-bottom: -50%;
+    background: #fff;
+    border-radius: 5px;
+    -webkit-box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);
+}
+img{
+    width:40%;
+    margin-bottom:10%;
+}
+.btnForget{
+    background: #c0392b;
+    border:none;
+}
+.forget-password .dropdown{
+    width: 100%;
+    border: 1px solid #ced4da;
+    border-radius: .25rem;
+}
+.forget-password .dropdown button{
+    width: 100%;
+}
+.forget-password .dropdown ul{
+    width: 100%;
 }
