+++ views/index.hbs
@@ -1,115 +1,2 @@
-<div class="carrusel">
-    <div class="row">
-        <div class="col-md-12">
-
-            <div id="carouselExampleControls" class="carousel slide" data-ride="carousel">
-                <div class="carousel-inner">
-                    <div class="carousel-item active">
-                        <img class="d-block w-100" src="http://www.miau.com.mx/wp-content/uploads/2016/08/campa%C3%B1a-oht.png" alt="Esterilizaciones">
-                    </div>
-                    <div class="carousel-item">
-                        <img class="d-block w-100" src="https://www.spca.nz/Images/Assets/80131/18/" alt="Third slide">
-                    </div>
-                    <div class="carousel-item">
-                        <img class="d-block w-100" src="http://salud.yucatan.gob.mx/wp-content/uploads/2013/11/slide_1campa%C3%B1adeesterilizacion.jpg" alt="Second slide">
-                    </div>
-                    <div class="carousel-item">
-                        <img class="d-block w-100" src="https://sahumane.org/images/adoption/2018/BFDayCFFL.jpg" alt="Third slide">
-                    </div>
-                    <div class="carousel-item">
-                        <img class="d-block w-100" src="http://www.amigosproanimal.org/inicio/wp-content/uploads/2019/01/%C2%A1ESTERILIZA-A-TU-MASCOTA.png" alt="Third slide">
-                    </div>
-                    <div class="carousel-item">
-                        <img class="d-block w-100" src="https://local.mx/wp-content/uploads/2019/02/perros-locales-perez.jpg" alt="Third slide">
-                    </div>
-                    <div class="carousel-item">
-                        <img class="d-block w-100" src="http://salemcountyhumanesociety.org/images/banner4.jpg" alt="Third slide">
-                    </div>
-                </div>
-                <a class="carousel-control-prev" href="#carouselExampleControls" role="button" data-slide="prev">
-                    <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-                    <span class="sr-only">Previous</span>
-                </a>
-                <a class="carousel-control-next" href="#carouselExampleControls" role="button" data-slide="next">
-                    <span class="carousel-control-next-icon" aria-hidden="true"></span>
-                    <span class="sr-only">Next</span>
-                </a>
-            </div>
-
-        </div>
-    </div>
-</div>
-<div class="container">
-    <div class="card-deck">
-        <div class="card">
-            <img class="card-img-top" src="https://bumerangnews.com/wp-content/uploads/2016/12/refugio-canino.jpg" alt="Card image cap">
-            <div class="card-body">
-                <h5 class="card-title">Card title</h5>
-                <p class="card-text">This is a wider card with supporting text below as a natural lead-in to additional content. This content is a little bit longer.</p>
-            </div>
-            <div class="card-footer">
-                <a href="/albergues"><small class="text-muted">Conoce más...</small></a>
-            </div>
-        </div>
-        <div class="card">
-            <img class="card-img-top" src="https://vetstreet.brightspotcdn.com/dims4/default/8810ede/2147483647/thumbnail/590x420/quality/90/?url=https%3A%2F%2Fvetstreet-brightspot.s3.amazonaws.com%2F56%2Fab%2F5e68cf5d4f2787e57dbe72027124%2Fdog-adoption-two-dogs-in-shelter-kennel-thinkstock-507654075.jpg" alt="Card image cap">
-            <div class="card-body">
-                <h5 class="card-title">Card title</h5>
-                <p class="card-text">This card has supporting text below as a natural lead-in to additional content.</p>
-            </div>
-            <div class="card-footer">
-                <a href="/adoptables">  <small class="text-muted">Conoce más...</small></a>
-            </div>
-        </div>
-        <div class="card">
-            <img class="card-img-top" src="https://d3u03kk87rjfaq.cloudfront.net/wp-content/uploads/2019/07/16112145/Dog-Cat.jpg" alt="Card image cap">
-            <div class="card-body">
-                <h5 class="card-title">Card title</h5>
-                <p class="card-text">This is a wider card with supporting text below as a natural lead-in to additional content. This card has even longer content than the first to show that equal height action.</p>
-            </div>
-            <div class="card-footer">
-                <a href="/adopta"><small class="text-muted">Conoce más...</small></a>
-            </div>
-        </div>
-    </div>
-</div>
-<div class="carrusel-adoptables">
-    <div class="row">
-        <div class="col-md-12">
-
-            <div id="carouselExampleControls" class="carousel slide" data-ride="carousel">
-                <div class="carousel-inner">
-                    <div class="carousel-item active">
-                        <img class="d-block w-100" src="https://miamidiario.com/wp-content/uploads/perros-en-refugio-890x395_c.jpg" alt="adoptables">
-                    </div>
-                    <div class="carousel-item">
-                        <img class="d-block w-100" src="https://durangopress.com/wp-content/uploads/2016/02/88.jpg" alt="adoptables">
-
-                    </div>
-                    <div class="carousel-item">
-                        <img class="d-block w-100" src="http://www.adoptacatfoundation.org/images/adoptions-main.jpg" alt="adoptables">
-                    </div>
-                    <div class="carousel-item">
-                        <img class="d-block w-100" src="https://local.mx/wp-content/uploads/2019/02/perros-locales-perez.jpg" alt="Third slide">
-                    </div>
-                    <div class="carousel-item">
-                        <img class="d-block w-100" src="https://images2.minutemediacdn.com/image/upload/c_crop,h_843,w_1500,x_0,y_78/f_auto,q_auto,w_1100/v1555182242/shape/mentalfloss/501389-istock-507640348.jpg" alt="adoptables">
-                    </div>
-                    <div class="carousel-item">
-                        <img class="d-block w-100" src="https://assets3.thrillist.com/v1/image/2547482/size/tmg-facebook_social.jpg" alt="adoptables">
-                    </div>
-
-                </div>
-                <a class="carousel-control-prev" href="#carouselExampleControls" role="button" data-slide="prev">
-                    <span class="carousel-control-prev-icon" aria-hidden="true"></span>
-                    <span class="sr-only">Previous</span>
-                </a>
-                <a class="carousel-control-next" href="#carouselExampleControls" role="button" data-slide="next">
-                    <span class="carousel-control-next-icon" aria-hidden="true"></span>
-                    <span class="sr-only">Next</span>
-                </a>
-            </div>
-
-        </div>
-    </div>
-</div>
+<h1>{{title}}</h1>
+<p>Welcome to {{title}}</p>
