+++ views/news/news.hbs
